+++ hardpy/common/stand_cloud/__init__.py
@@ -1,0 +1,13 @@
+# Copyright (c) 2024 Everypin
+# GNU General Public License v3.0 (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
+
+from hardpy.common.stand_cloud.connector import StandCloudConnector
+from hardpy.common.stand_cloud.exception import StandCloudError
+from hardpy.common.stand_cloud.registration import login, logout
+
+__all__ = [
+    "StandCloudConnector",
+    "StandCloudError",
+    "login",
+    "logout",
+]